+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - fix route policy `standard_community_lists` to properly handle both single and multiple community list entries
 - add support for classic policy object port list
 - add support for service LAN VPN Ethernet Interface
+- add support for service LAN VPN GRE Interface
 - fix `trunk_allowed_vlans_variable` not being set correctly in switchport feature template
 
 ## 1.2.0